fix: Refactored show method inside LogParserView and moved to YGuardLogParser using "Move method/field" method.
